refactor: Text fields and colors selectors migrated to another new components
Francisco Zambrano
