+++ src/components/Generator/ColorSelectors/index.js
@@ -1,0 +1,1 @@
+export {default} from './ColorSelectors';
