+++ src/components/Generator/Generator.scss
@@ -1,0 +1,112 @@
+.button-download {
+  &__desktop {
+    width: 200px;
+    margin: 0 auto;
+
+    @media (max-width: 767px) {
+      display: none !important;
+    }
+  }
+
+  &__mobile {
+    position: fixed;
+    bottom: 1.4em;
+    right: 1em;
+    opacity: 1;
+
+    &:active {
+      opacity: 0.8;
+    }
+
+    svg {
+      width: 4.2em;
+      height: 4.2em;
+    }
+
+    @media (min-width: 768px) {
+      display: none;
+    }
+  }
+}
+
+.logoPacto {
+  width: 8em;
+  margin: 2em auto 3em;
+}
+
+.hashtag {
+  color: #AE43A1;
+  display: flex;
+  justify-content: center;
+  font-weight: bold;
+}
+
+.letter-box {
+  margin-top: 2.5em;
+  display: flex;
+  justify-content: center;
+
+  .letter {
+    font-size: 5em;
+    font-family: 'RubikOne', serif;
+    text-transform: uppercase;
+    display: inline-block;
+    opacity: 0.7;
+    width: 0.50em;
+
+    &-i {
+      width: 0.26em;
+      padding-left: 5px;
+    }
+
+    &-w, &-m {
+      width: 57px;
+    }
+
+    &-c {
+      width: 47px;
+    }
+
+    &-0 {
+      color: #ff9400;
+    }
+
+    &-1 {
+      color: #00208b;
+    }
+
+    &-2 {
+      color: #ff001b;
+    }
+
+    &-3 {
+      color: #14ae40;
+    }
+
+    &-4 {
+      color: #b10a9d;
+    }
+  }
+
+  .logo {
+    margin-top: 0.5em;
+    display: inline-block;
+
+    svg {
+      width: 4em;
+      opacity: 0.7;
+      transform: translate(-0.15em, -0.1em);
+    }
+  }
+
+  .loguito {
+    display: inline-block;
+    transform: translate(-0.4em, -0.9em);
+
+    svg {
+      width: 3.8em;
+      opacity: 0.7;
+    }
+  }
+}
+
