+++ src/components/Generator/TextFields/TextFields.scss
@@ -1,0 +1,6 @@
+.text-fields {
+  display: flex;
+  justify-content: center;
+  flex-direction: row;
+  margin-bottom: 1em;
+}
