+++ src/components/Generator/TextFields/index.js
@@ -1,0 +1,1 @@
+export {default} from './TextFields';
